updated base page and product page
