+++ tests/ui/Product_Display_Page/TC_PDP_029_Validate_navigating_to_PDP_using_Product_Image_in_Cart_button_toggle_box_test.py
@@ -42,10 +42,10 @@ def test_validate_navigating_to_pdp_using_product_image_in_cart_button_toggle_bo
     expect(product_page.get_confirmation_message()).to_be_visible()
 
     # Step 5: Click on 'Cart' button which is in black color beside the search icon button on the top of the page
-    home_page.click_cart_button()
+    product_page.click_cart_button()
 
     # Step 6 (ER-1): Click on the Product Image in the displayed toggle box
-    new_product_page = home_page.click_cart_image_link()
+    new_product_page = product_page.click_cart_image_link()
 
     # Validation: User should be taken to the Product Display page of the Product
     expect(new_product_page.lbl_product_name).to_be_visible()
